requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.6.1
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
tiktoken>=0.5.2 